Move RecordSelector to components subpackage
