+++ src/main/resources/com/turtlesamigo/controllers/sample-overview.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.Pane?>
-<?import com.turtlesamigo.controllers.RecordSelector?>
+<?import com.turtlesamigo.controllers.components.RecordSelector?>
 
 <AnchorPane minHeight="1024" minWidth="2048" prefHeight="1024" prefWidth="2048" xmlns="http://javafx.com/javafx/11.0.1"
             xmlns:fx="http://javafx.com/fxml/1"
